src/tests/requirements.txt
pytest
requests
kafka-python
pytest-asyncio >= 0.21
